backend/requirements.txt
Django==2.1.5
psycopg2==2.8
djangorestframework==3.9.2
markdown==3.1
django-filter==2.1.0
